finish 最小生成树的扩展应用

+++ codeforce/1076(div3)/D. Monster Game.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 const int N = 2e5 + 10;
 
-int a[N], b[N];
+long long a[N], b[N];
 int n;
 
 void solve()
